Dispatch launch only after search result loads
We move the dispatch of the first-launch action into the callback for
fetching the search result, when launching from a search result.

Ideally we would do this for launching from search as well, but the mesh
interaction's do_search() method doesn't have a callback and we can't
easily add one without a large refactor.

[endlessm/eos-sdk#3795]