Show appropriate pages for all launch methods
The window module happens to default to showing the home page, because
that's the page that gets added first to its stack. This means that every
other page will transition in from the home page, and we don't want that.

Instead, make sure the window's very first transition has a duration of 0,
and dispatch an action to show the appropriate page as the very first
thing we do when the interaction launches.

[endlessm/eos-sdk#3795]