Don't reparent document card
In the article stack module, reparenting the document card was a holdover
from when we still used an offscreen window to render the webpage. It's
now impossible for the document card to have another parent, so we can go
ahead and add it to the article stack and show it.

For the latter, we have to make sure the webview is visible too.

(Matt W & Philip)

[endlessm/eos-sdk#3779]