Fetch suggested articles when requested from module
When the suggested articles module needs some suggested
articles, make a request to get a few random articles.
Need to ensure that a set of random articles is tied
to a specific search query so that even when going back
in the navigation history, you get the same 'random' results
for a given search.

[endlessm/eos-sdk#3822]