sideMenuTemplate: only trigger menu when mouse on left
Take two. Trigger the menu when event.get_coords().x < 3 will trigger
improperly when mousing over gdk windows that aren't on the left of
the template.

By using absolute coordinates everywhere in our calculations we should
not only trigger for mouse motion that really on the left three pixels
of the template
[endlessm/eos-sdk#3801]