articleStackModule: show card immediately with none transition
When we load an article with loading animation NONE, that means we want
to show it immediately instead of waiting for load-committed. This fixes
the problem where you would transition to an article page that was still
showing a glimpse of the previous article, while the next article was
still loading.

(Matt W & Philip)

[endlessm/eos-sdk#3779]