Dispatch first launch when sets are loaded
Previously we would dispatch the "first launch" action when the app was
launched for the first time, as its name suggests. This is still what we
do for launches from search and from search results. For desktop launch,
however, we need to dispatch first launch only after the sets have
finished loading, so that we never show an unpopulated home page.

This involves splitting the initial sets query out into its own async
method, and calling that method in the launcher methods instead of in the
constructor. For desktop launch, we then dispatch the appropriate actions
in the callback. For the other two launch types, the callback is just a
dummy callback, because we don't start out showing the home page, so the
sets can just load in the background.

[endlessm/eos-sdk#3795]