aisleInteraction: hack search results model page number
We have a bug where all reader app search results cards show up as
archived. We used to set the cards page-number property when creating
the card from the interaction, but now the card is created from the
search module, which has no way of knowing about the application
state needed to calculate page number.

For now we will do an aweful hack where we reset the models page
number to be the user facing page number. In the future we will
need to figure out a better way for the ReaderCard to access that
application state.
[endlessm/eos-sdk#3829]