Render borders, triangles, and arrows on cards
This (ab)uses the "outline" CSS property and gtk_render_focus() to render
a blue border on top of the cards on hover, hopefully without causing a
reallocation, which the "border" property would do.

It also renders triangles with arrows inside them in the lower right
corners. I created the arrow SVGs myself since the one from design was
actually a PNG.

InVision specs:
https://endlessm.invisionapp.com/d/main#/console/4678196/107218154/preview

[endlessm/eos-sdk#3797]