Add context label to post cards
Similar to the thumb cards, the context
label here shows a subset of tags assigned
to this article model. In the case
of post cards, we have to make sure this
label only shows up for post cards which
are backed by an article model, rather than
a set model.

[endlessm/eos-sdk#3786]