Fixes to Template B apps on composite mode
We add some visual fixes to Template B apps so that they are displayed nicely
on composite mode, which calls for a 720x480 resolution.

The biggest change is the adjustments required by the sidebar template to reduce
the width of the sidebar when the app window is reduced below a certain
threshold.

[endlessm/eos-sdk#3752]