Process dispatcher actions only when needed
Instead of checking the dispatcher queue in an idle function that runs
over and over again, only start processing the queue when an action is
dispatched. Also stop the idle function when the queue becomes empty.

This prevents CPU load from going to 100% while the app is not doing
anything.

This makes the dispatcher's start() method obsolete, so remove it.

[endlessm/eos-sdk#3833]