thematicModule: refactor visibility logic
Rather than having a way to map arrangement to header and carefully
updating header visibility, we add both the header and arrangement
to a grid and bind the visibility of that grid to the visibility of
the arrangement.

This will make it easier to add a separator widget which should also
be hidden.
[endlessm/eos-sdk#3861]