All apps logos can fit in composite mode
Some of our knowledge apps were not being properly displayed in composite mode
due to large logos that were not being able to size down to fit this resolution,
so we need to anticipate that.

The unfortunate side effect is that in larger resolutions, the apps can be sized
down to really small dimensions, causing the logos to look tiny! This is a
tradeoff that we must take for now, until we redesign the bad-looking logos or
take other measures as resizing cards in the home pages.

[endlessm/eos-sdk#3827]