Archive notice label can span two lines
In some cases in which the reader app title is too long, the archive notice
label may push the width of the whole app to be over 720px wide, which is our
target resolution for composite mode.

Letting this label span over two lines fixes that. It may look ugly in some
cases, but it is a reasonable treadeoff to achieve composite resolution.

[endlessm/eos-sdk#3827]