Don't tie showing brand screen with first launch
We don't always want to show the brand screen on
launch, e.g. if we are launching due to a desktop
search. So we shouldnt tie those two actions together.
The interaction (presenter) should be responsible
for separating the logic between those, not the window.
To this end, we create a new action: SHOW_BRAND_SCREEN,
which gives the interaction more fine grained control
over when to show the brand screen.

[endlessm/eos-sdk#3862]