Restore cards fading in on preset A
Previously we faded in cards in the search results and set group in
preset A with a special case in the presenter. When we took out the
special case, we forgot to add the fade-in back in elsewhere. This
restores it.

For this, we add a CardContainer interface. A fade-cards property is the
only thing it currently has, but there are opportunities for more common
code between the modules that implement it.

[endlessm/eos-sdk#3156]