Don't add a CSS provider in size-allocate
We can set the background size in CSS to the correct value using
percentages, which makes it unnecessary to add a CSS provider in every
size allocate, and also unnecessary to load the background image in a
pixbuf in order to determine its size.

This should make things run a bit more efficiently.

[endlessm/eos-sdk#3590]