highlightsModule: refactor set loading
This is a big shift, changes things so that each arrangement populates
with a separate request to the engine, so that later we can limit the
number of models we fetch for each arrangement.

As part of this work, changed how the featured arrangement is loaded.
Instead of showing a combination of cards from the other two sets we
are showing, we will just show another featured set entirely. This
is both simpler to implement and avoids the problem of showing
duplicate cards on the home page. Still totally faking what eventually
should be featured cards though.
[endlessm/eos-sdk#3571]