Add WebviewTooltip management to the Travel App
The new Travel app adds WebviewTooltips to URLs in the article pages. We add the
ability to handle these tooltips through the WebviewTooltipPresenter.

A new, default tooltip type is added to the tooltips template file. This tooltip
type only shows a single label, which in the case of the Travel app, is the text
of the URL for which the tooltip appears.

[endlessm/eos-sdk#3799]