Code branch on whether content model is set or not
Previously we assumed that when a search result card
was clicked, it was always an article model that was
backing it, so we would try to transition to the
article page. However, sometimes those cards are backed
by set models. We should check which one it is and
act accordingly.

[endlessm/eos-sdk#3863]