Make sure synopsis label on search result card is always visible
Some cards, most obviously those backed by set models,
do not have synopsis. Normally in the card interface,
when no synopsis is given, the synopsis label
is hidden. In the case of search result cards
however, we dont want this to happen since it will mess
up the layout of widgets on the card. So lets force
it to always be visible.

[endlessm/eos-sdk#3863]