Make context banner label part of history item
In order to ensure that the label on the context banner
is always relevant to the article currently being viewed
we need to add this context label to the history item
object.

It can get set by any module which hosts an arrangement
of card and wants to customize the context label above
those cards.

[endlessm/eos-sdk#3794]