Lazy load content on the Thematic module
Previously the thematic module was taking > 5 seconds
to load, since it had to create models and cards
for entire sets worth of content. Now instead we load
one arrangement at a time and only add a new arrangement
when the user scrolls to the bottom of the page.
In order to do this, we have a new method on the
module interface, which is show_more_content. If
a module has a 'scroll-server' slot, then this
method will get invoked once the scroll reaches the
bottom of the page.

[endlessm/eos-sdk#3884]
[endlessm/eos-sdk#3884_prime]