Lazy load content on the Thematic module
Previously the thematic module was taking > 5 seconds
to load, since it had to create models and cards
for entire sets worth of content. Now instead we load
one arrangement at a time and only add a new arrangement
when the user scrolls to the bottom of the page.

In order to do this, we add a new 'scrollable' interface.
The init method of this interface has the module listen
to dispatches from its scroll server. A scroll server
is a scrolling template module which sits above the
scrollable module in the widget hierarchy. When the
scroll server decides it needs more content, it dispatches
a NEED_MORE_CONTENT signal, providing its own unique
name in the scroll-server field of the payload. The scrollable
module will listen to this dispatch, and, if the scroll-server
field matches its own scroll-server property, it will
add more content.

[endlessm/eos-sdk#3884]