Standardize names of payload properties
We were inconsistent about using 'text' or
'query' as the name of the string passed from
the search box to the interaction module.
This standardizes it to always be 'query'.

Note that we keep the name 'text' to refer
to when the interaction module sets the string
on the search box.

[endlessm/eos-sdk#3931]