Fix buffet search result cards on composite
Giving them a minimum width class of E doesn't quite fit on the screen,
so we need to change the minimum to D and add styling for that in the
normal case. We also add some padding to the list box rows on composite
so that the cards are not squished up against the edge.

[endlessm/eos-sdk#3933]