Change UI font to Roboto on composite
Lato at smaller sizes is not a good font for composite TVs. Replace all
of the normal-size text in the A, B, Reader, and Encyclopedia apps with
Roboto on composite. We leave the buffet apps as they are because the
text is all large enough to be legible there.

We can also remove * { font-family: Lato; } from everywhere, because the
theme already does that for us.

[endlessm/eos-sdk#3933]

Conflicts:
	data/css/endless_buffet.css