css: remove card b text shadow and box shadow
The text shadow is white on white background, you can't really see it
and its expensive to compute.

The box shadow, similarly was black on a black border. You can only
just see it leaking onto the other card images, which is quite odd
anyway.

Let's ditch both of these as they are questionable and a big
performance hit.
[endlessm/eos-sdk#3942]