meshInteraction: dispatch first launch quickly
When loading an article directly, we show dispatch SHOW_ARTICLE_PAGE
immediately to show a blank article page to the user.

This is commented and seems quite intentional, so we should make
sure to also dispatch first launch right away so the window is
displayed to the user immediately as well.
[endlessm/eos-sdk#3864]