application: don't null interaction on window remove
First off, this was scary code that seemed quite breakable, if
*any* window associated with the application was removed, even if
its not the main window, we would destroy all our application state.

Secondly, with the inactivity timeout remove this should no longer be
necessary. There are two application launching code paths

If the application is launched with --gapplication-service the app
will wait for 10 seconds for any of its dbus methods. On method call
the gtk will put a hold on the application and the timeout will be
removed.

If the application is launched without the service flag, activate
will be called right away and gtk will hold the application.

In both cases gtk will release the window immediately after the
main window is destroyed and the application will shutdown immediately.
So basically, we should never have to create the main window twice on
a single run on the application.
[endlessm/eos-sdk#3864]