Add tiny margin to responsive margins module
For small window sizes (<800 px) we add another tier of margins to the
responsive margins module. This is necessary because the margin tier that
was previously smallest (40px on each side) was causing apps to be larger
than the screen width on composite TVs.

The new smallest tier is called "tiny" and currently removes the margins
altogether.

[endlessm/eos-sdk#3973]