Justify app subtitle according to banner alignment
In order to look good when the subtitle label wraps, we need to set its
justification according to the app banner's horizontal alignment.

This centers the subtitle under the app logo on e.g. the brand screen of
the travel app, and left-justifies it on e.g. the home screen of the
reader apps.

[endlessm/eos-sdk#3973]