Limit synopsis label to an integer number of lines
Previously the synopsis label on Card A would extend its allocation off
the bottom of the card, and also render partial lines if the card
happened to be clipped in the middle of a line of text.

This sets the synopsis label's lines property to the largest integer
number of lines that fit inside the card's allocation. We set the lines
property in an idle function so as not to queue a resize from within
size_allocate(). We also make sure to do it only once per card.

This has a few drawbacks: on composite, there is a short but visible
delay before the text is formatted correctly. And, this code probably has
a small performance impact since it does require an extra round of
resizing for each card.

This allows removing the workaround to lower the number of lines in
composite mode.

[endlessm/eos-sdk#4005]