Implement removing cards from arrangements
Previously we could only remove all cards at once. This makes
arrangements conform more to the expectations of a Gtk.Container by
allowing individual cards to be removed. I'm not sure whether we'll need
this in the actual apps, but it will allow the testing tool to have a
remove button which really helps in understanding how the cards should
resize and reflow in a particular arrangement.

For the arrangements which proxy another container (Grid, List), this
requires monkeypatching remove(), because of an assert in
gtk_container_remove() that checks whether the widget you are removing
is really a child of the parent, which prevents you from having an
internal container inside a container widget.

(GTK itself does have internal containers inside container widgets, e.g.
GtkAssistant, but every time they do that, they add a check to that
assert to allow it for that specific widget ... sigh.)

[endlessm/eos-sdk#3771]