Picard, the arrangement smoke tester
This is a utility that will help us smoke-test our arrangements, and
hopefully also help the design team to understand how cards and
arrangements interact, resize, and reflow.

Run "picard" with no arguments to print a list of the arrangements that
are available to test. Then run it with the name of the arrangement, e.g.
"picard SquareGuys".

Click the "+" button to add a card to the arrangement, the "-" button to
remove the last card added, and the trash button to clear away all cards.
The menu allows you to change the spacing between cards, or resize the
window to a particular screen size.

[endlessm/eos-sdk#3771]