Add a link to completion popup to see more results
The number of results displayed in the SearchBox auto-completion
popup is limited to REQUEST_SIZE(10). The user has no idea about
the all other results.

Therefore add a link to popup using a completion action, so the
user can see more results than displayed in the popup, just as
if they pressed ENTER in the search entry.

Add initial style to the link, using common css and more specifics
in endless_buffet css.

[endlessm/eos-sdk#3717]