Fix test relying on faulty equality
In the following commit we ensure that all models have an EKN ID. This
broke a test in HighlightsModule where two models were erroneously
comparing as equal by Jasmine because they didn't have an EKN ID and
otherwise had the same properties. This fixes the test so that it checks
that the model present in the payload is the model belonging to the card
that we triggered the click on.

In addition, mock out the Utils.shuffle function before anything else,
which is good practice since it could be called in the constructor of
HighlightsModule.

[endlessm/eos-sdk#4053]