Put context label inside space container
Previously, the context label was awkwardly
spilling over onto two lines in composite
mode. This fixes that by splitting them into
two labels, and putting both in a space
container, thus ensuring that if there is
not enough room to show both labels,
only one will get shown.

We also sort context labels from shortest
to longest in length to maximise the chances
we get to show two of them.

[endlessm/eos-sdk#4017]