spaceContainer: properly request height for width
And width for height. Before we were only allocating with height
for width but not requesting sizes with it.

There's still a problem though when a parent container ignores the
preferred request mode and just tries to do constant size anyway.

In that case, because the space container always allocates width
for height of height for width anyway, it can still end up rendering
empty.
[endlessm/eos-sdk#4037]