Preserve original spacing in DividedBannerTemplate
The original row spacing was being overridden with smaller and
smaller values each time it allocated its size, when the width
available was less than 800px.

Store the original row spacing value only once, the first time
size_allocate is called.

Add tests for preserving this behavior and to guarantee that this
function is idempotent.

[endlessm/eos-sdk#4056]