Allow registering local types with warehouse
This allows programs such as Picard to register their own types with the
warehouse, so that they can be referred to from app.jsons. We use it here
for ColorBox, but it will be quite useful when porting the other smoke
tests to use app.jsons.

Eventually we'll need to allow apps to supply custom modules that aren't
in our main tree, and this is the basis of that as well.

[endlessm/eos-sdk#4000]