window: put search bar in a stack and leave visible
This accomplished two things. First, the top bar will always be constant
sized. This actually comes into play with composite mode with the text
scaling factor. Second, showing the search bar will only queue a allocate
and draw on the stack itself, rather that a draw on the whole window.
[endlessm/eos-sdk#3590]