Add card-type slot to ReaderWindow
Part 1 of moving the card-type slot off of AisleInteraction. ReaderWindow
now has the slot that determines the type of document cards. However,
AisleInteraction still has a card-type slot that determines the type of
document cards on the standalone page. That will be removed in the next
commit.

Note that the card-type slot will not stay on ReaderWindow for too long
either, as it should get moved to the document-arrangement.

[endlessm/eos-sdk#4059]