Move document card slot to standalone page
This is part 2 of removing the document card slot from AisleInteraction.
In the previous commit we moved half of that responsibility onto
ReaderWindow, in preparation for later moving it into an arrangement.
This commit moves the other half of that responsibility, creating
document cards for the standalone page, onto StandalonePage itself.

[endlessm/eos-sdk#4059]