Simplify progress label
Make the progress label a permanent part of the reader document card,
instead of passing it in at construct time. Remove the info-notice
property from ReaderDocumentCard and instead add four properties that
control what is displayed in the progress label.

These properties are out of place and still not very modular, since the
ReaderDocumentCard can still only work with the ReaderWindow and the
StandalonePage. However, this is a start in the direction of
endlessm/eos-sdk#4036, where card models should gain awareness of their
place in the collection in which they are displayed.

We depart slightly from the design specs and don't specify the magazine
title if you click on a link to an archived article. Presumably since you
started up the app yourself, you know what the title is. We still do
display the magazine title when reaching the standalone page from global
search, since it's not so clear in that case. This simplifies the code
quite a lot and allows better compartmentalization between modules.

[endlessm/eos-sdk#4059]