Setting compat props on module type always applies
When setting module properties from v1 app.jsons in the compatibility
layer, we have a big switch statement that handles various named modules
in the module hierarchy. The goal is eventually to get rid of this
switch statement and set properties on all modules by type, so that they
function correctly no matter where they are put.

To this end, we can move calls to set_prop_for_type() outside the switch
statement.

[endlessm/eos-sdk#4059]