Dispatchify some things around standalone page
When viewing an archive article, the standalone page is displayed.
Additionally if the request is coming from global search, it shows a
banner with the app background and a button that says "open
application".

Previously the interaction reached into the standalone page's internals
in order to connect to the button's signal. Instead we add a dispatcher
action, "leave preview clicked".

Also previously, the interaction did not dispatch "show article" in this
case, and loaded the article by directly calling a method on the
standalone page. We also want to get rid of that, because the standalone
page needs to be modular.

[endlessm/eos-sdk#4059]