Add supplementary articles module.
Add a module that will show a number of supplementary
articles. The same-set property allows you to
specify if these articles should be part
of the same category as the current object model
or from different ones.

It tries first to show unread articles. If no unread
articles are available, it shows read articles.

[endlessm/eos-sdk#3995]