Split standalone page and archive page
The goal of this large refactor was to bring StandalonePage into the
modular system, something we never previously had a chance to do. It's
necessary now for issue endlesssm/eos-sdk#4000, because we are
refactoring the card-type slots, and StandalonePage has a card-type slot.

StandalonePage was previously used in one place in the reader app,
but for two functions:
  1) displaying an article with a fancy banner above it, when you click a
     result from global search that is in the archive.
  2) displaying a normal article page with a message saying that the
     article is in the archive, when you reach the article from a link in
     another article, or from search results.

Henceforth, to avoid confusion, #2 is referred to as the "archive page".

It becomes convenient to split the two uses of StandalonePage in app.json
to avoid a lot of "if from_global_search" logic. By doing this, the fancy
banner is now split off into its own module, StandaloneBanner.

ReaderWindow gains an extra slot, "article-page", as distinct from
"standalone-page". It loses the "standalone-page" public property, which
was a long-standing FIXME.

There is one slight change in user-facing functionality. The archive page
does not display the name of the app in its message any longer. The
reasoning behind this is that the user needs that context on the
standalone page since it's the first thing they see when they open the
app, but on the archive page they already know what app they're in.

P.S. To test the standalone page, run

gdbus --session -d com.endlessm.domain-lang -o /com/endlessm/domain_lang \
      -m com.endlessm.KnowledgeSearch.LoadPage 'ekn://domain-lang/sha' \
      'some text' 0

If you are using jhbuild, make sure the app is already running using your
modified eos-knowledge-lib so that a new copy is not started up using the
old eos-knowledge-lib.

[endlessm/eos-sdk#4059]