Separate standalone article dispatcher event
Previously, show article was used for showing articles on the standalone
page as well as in normal use. This meant that we had to keep passing
around a "from_global_search" parameter on the event payload. Since the
two usages are semantically different, we may as well have two different
events.

[endlessm/eos-sdk#4059]