QuarterArrangement
The QuarterArrangement is an set arrangement that displays a single row of
featured cards and as many rows of support cards as needed.

The featured row may show between two and four featured cards, depending on the
width of the arrangement. The support rows may show between two and three cards
per row, again depending on the width of the arrangement.

The arrangement has three width stop points to arrange cards:
 - 1200px: Four featured cards and three support cards per row are shown.
 - 900px: Three featured cards and two support cards per row are shown.
 - 720px: Two featured cards and two support cards per row are shown.

[endlessm/eos-sdk#3986]