Avoid repeated categories between modules
Categories displayed in highlights module can also be seen in
other modules, e.g. suggested categories module in travel app.

This happens because all modules are fed with the exact same set
of categories, so it is likely that they will display the same ones
because it is random.

This approach adds a new signal called FILTER_SETS to allow modules,
such as highlights module, to broadcast the IDs of the categories they
are displaying. Other modules can listen to this signal and remove
categories if necessary.

[endlessm/eos-sdk#4035]