Use TopbarHomeButton in A B and Buffet templates
As agreed with the design team, all apps should include a generic home
button at the top left corner of the app window. This button must take
the user back to the home page.

Add this new generic home button to EknWindow to ensure it is present in
all knowledge apps based on templates A, B and Buffet.

Listen to HOME_CLICKED signal in MeshInteraction to do the transition to
the home page through the HistoryPresenter logic.

[endlessm/eos-sdk#4030]