Refactor arrangements to take models
Previously arrangements were given card widgets. Instead, give content
object models to arrangements, and have them create the cards themselves
from the models. This allows us to start work on filtering and sorting in
arrangements. It also opens up the possibility to avoid creating card
widgets in cases where the widget isn't shown in the arrangement,
although this commit performs no such optimization.

The add_card() API and friends are removed from Arrangement, and replaced
by add_model(), remove_model(). Arrangement implementations are now
required to be GtkContainers. For containers that require more
complicated packing than just add() and remove(), implementations can
override pack_card() and unpack_card(). These two are meant to be for
implementations only, not to be called externally. They could be replaced
by vfunc_add() and vfunc_remove() if it weren't for the fact that add()
and remove() assert that the child is a direct child of the widget.

There are some concessions to convenience, such as get_card_for_model()
and the fact that add_model() and remove_model() return the corresponding
card. I'm not convinced this is necessary and will attempt to simplify
the API in a following commit, but this commit forms a baseline.

This moves all card-type slots from card containers to arrangements,
which also requires some rearrangement of the compatibility app.json
files.

There is one regression; search words are not highlighted on cards in the
travel app. That will be fixed in a following commit.

[endlessm/eos-sdk#4000]