Repair search term highlighting on cards
This repairs the functionality that I broke in the previous commit,
namely that of search terms being highlighted on search result cards in
the travel app.

Since we want to avoid having any other modules besides arrangements deal
with card widgets, we move search string highlighting into the
arrangements. Previously, cards could only be highlighted at construction
time, but it doesn't seem wise to guarantee that anymore now. So this
requires some changes and cleanup in the Card interface as well: e.g.,
Card.set_title_from_model() should not do highlighting, so that cards
that shouldn't support search term highlighting, such as CardB, never get
highlighted.

A bonus of this refactor is that it is now trivial to enable search
term highlighting on CardA, with a one-line change to MeshInteraction.

[endlessm/eos-sdk#4000]