Share card-fading responsibility
This functionality has lived in several different places now, but I think
this now makes the most sense. We don't want anything outside of
arrangements to deal with actual card widgets, so arrangements must be
responsible for deciding whether to call fade_in() or show_all() on
cards. However, sometimes whether to fade in is determined semantically
by the card container that controls the arrangement, e.g. don't fade in
the first batch of cards but do fade in subsequent batches.

Sometimes cards should not fade in because the arrangement is not
designed in such a way to support that, for example CarouselArrangement;
in that case, the CarouselArrangement overrides an
Arrangement.fade_card_in() method (effectively a vfunc) to not fade cards
in.

Sometimes cards should not fade in for semantic reasons, and in that case
the card container decides so by setting the Arrangement.fade_cards
property.

[endlessm/eos-sdk#4000]