Proxy card clicked signals on Arrangement
This almost completes the move of responsibility for card widgets into
the arrangements. Previously card containers would take the card returned
by add_model() and connect to its 'clicked' signal. Since only
arrangements should deal with card widgets, we instead connect to that
signal inside the arrangement, and proxy it as the 'card-clicked' signal.
The 'card-clicked' signal sends along the card model as a parameter, not
the card widget itself.

There are two inconveniences left over in the new division of
responsibilities:

1) DocumentCards don't have a 'clicked' signal, because they're not
   GtkButtons. We should consider also making them not cards, either, and
   not sticking them in arrangements, but that's for the future.

2) In ReaderWindow we still need to look up the DocumentCard with
   get_card_for_model() in order to connect to the 'ekn-link-clicked'
   signal. ReaderWindow is scheduled for demolition in the future, so
   we can leave it like this for now.

[endlessm/eos-sdk#4000]