Add margin to compensate top menu
When the top menu is open, it covers the top of whatever is in the
content slot. To compensate, this adds a top margin to the content slot.
A complicated case is when the content slot contains a scrolled window,
with optional viewport. In that case, the child of the scrolled window or
the viewport (if present) needs to get the extra margin.

[endlessm/eos-sdk#3992]