Add AllTypeCard
Card for displaying articles in the News App. This widget can
display articles using labels for the title, the synopsis and
the context which is a tag, if there is any available.

Also, add a new css specifically for the News App.

This card requires Fira Sans and Merriweather fonts.

[endlessm/eos-sdk#4098]