Get rid of explicitly nulled slots
Until now our design decision was to make all slots mandatory, and
indicate the choice to leave a slot empty by a value of null for that
slot.

That is not backwards-compatible if you add new slots. Also, on this
issue we're going to need to add a bunch of slots, so now is a good time
to make this change.

For now, optional slots are indicated only by a comment. Hopefully in the
future we can have some kind of programmatic indication that a slot is
optional.

[endlessm/eos-sdk#3990]