Extract common card-adding code
This can go into a function in minimal.js, since we'll want it later on
in this pull request to add cards that can be ordered the way we expect
them to. The idea is that we give the cards an easily sortable tag so
that we can predictably retrieve them in the same order that they are
returned from factory.get_created_named_mocks().

[endlessm/eos-sdk#3990]