ekn-history-model: handle extra go_back/go_forwards
This commit switches up the api of the history model to do nothing
when calling go_back with no previous items and the history and
similarly for go_forward. Before we would throw an error.

This fits better our current use case for the history mode, while
not ruling out use case (you could still throw an error using the
history model if can_go_back is false).
[endlessm/eos-sdk#4095]