Card ordering in arrangements
This adds an "order" slot to arrangements and an Order interface for
objects that can fit into that slot. The slot is optional.

If the slot is filled, then the arrangement will display its cards in
the order dictated by the Order object. (If it so chooses. The Carousel
arrangement currently ignores ordering.)

Implement an order by overriding its compare_impl() method. A sample
implementation, AlphabeticalOrder, is provided.

Card ordering is hard to test, since it's displayed differently for every
arrangement. In this commit, it's tested implicitly for most arrangements
by giving them an order (Minimal.CardCreateOrder which sorts cards in the
order in which they're created by Minimal.create_ordered_cards()) and
relying on that to be correct when we test the cards' width and height.

Arrangement.get_models() now always returns a sorted list of models as a
consequence of this change.

[endlessm/eos-sdk#3990]