Arrangement.get_count()
Since get_models() sorts an array now, this should provide a way to get
the count of models / cards in an arrangement that is quicker than either
get_models().length or get_children().length.

I would have liked to make this a property, but we currently do not
inherit property getters from interfaces.

[endlessm/eos-sdk#3990]