aisleInteraction: dispatch page updates to the window
This makes zero attempt to clean up the api between the window and
interaction. It just takes the calls that were direct and moves them
wholesale to the dispatcher.

This will fix a bug introduced by pausing the dispatcher while
animating. We want to dispatch SEARCH_STARTED before showing the
search page, but because dispatcher actions are shown in an idle
and the window show page methods were being called synchronously,
there was no way to do that.
[endlessm/eos-sdk#4095]