Support secondary sorting in order interface
There will be sorting classes in which most of the
models are going to be equal, e.g. when ordering by
featured property. This can be solved by supporting
a secondary sorting class.

Add a "sub-order" slot to the order interface so order
classes can fallback to a secondary sorting class when
two elements are equal.

[endlessm/eos-sdk#3997]