Use SetMap to get parent set titles for articles
Previously, on a card's context label, rather than
showing the parent set's title, we were just showing
the article's tags. These coincidentally happened to
be same as the parent set's title, but that may not
always be the case. Now, with the SetMap, we have
a way of getting the actual set title so let's do
that.

[endlessm/eos-sdk#4034]