Add DominantColor helper
Ideally the dominant color would be a property of
the object model but, to achieve that, it would
require to process these thumbnail images at build
time, for every model that is affected.

For now we have agreed that this should be done in
run time. One of the reasons is to avoid adding more
changes to building pipeline in this current window.

Therefore add a new helper called DominantColor which
provides a get_dominant_color function. This function
takes an <ContentObjectModel>, fetches the image model
if necessary and extracts the dominant color.

Note this is a temporary solution.

[endlessm/eos-sdk#4002]