Add BackgroundModule
The News App design requires that each page has its own
customized background. The basic customization consists
in placing a blurred image with a tinted overlay on top
of the image and a beige overlay to cover the bottom.

The overlay color is the dominant color from a featured
article.

Add BackgroundModule which receives a <ContentObjectModel>
by listening to the FEATURE_ITEM action and then it draws
the background as it is described above.

https://phabricator.endlessm.com/T10586